requirements.txt
bcrypt
mysql-connector-python
flask
flask_mail
dotenv